@@ -2,26 +2,30 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 apply plugin: 'kotlin-kapt'
+apply plugin: 'com.google.gms.google-services'
 
-configurations.all {
-    resolutionStrategy.eachDependency { DependencyResolveDetails details ->
-        def requested = details.requested
-        if (requested.group == 'com.android.support') {
-            if (!requested.name.startsWith("multidex")) {
-                details.useVersion '27.1.1'
-            }
-        }
-    }
-}
+//configurations.all {
+//    resolutionStrategy.eachDependency { DependencyResolveDetails details ->
+//        def requested = details.requested
+//        if (requested.group == 'com.android.support') {
+//            if (!requested.name.startsWith("multidex")) {
+//                details.useVersion '27.1.1'
+//            }
+//        }
+//    }
+//}
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.test.weatherdemo"
-        minSdkVersion 19
-        targetSdkVersion 27
+        minSdkVersion 23
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
     kotlinOptions { jvmTarget = 1.8 }
     buildTypes {
@@ -50,9 +54,6 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$rootProject.kotlin_version"
 
 
-    api "com.android.support:appcompat-v7:$rootProject.support_version"
-    api "com.android.support:recyclerview-v7:$rootProject.support_version"
-
     // Kotlin Library
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$rootProject.kotlin_version"
 
@@ -75,17 +76,21 @@ dependencies {
 
     //Dagger2
     api "com.google.dagger:dagger:$rootProject.dagger_version"
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
     kapt "com.google.dagger:dagger-compiler:$rootProject.dagger_version"
 
+    implementation platform('com.google.firebase:firebase-bom:30.0.1')
+    implementation 'com.google.firebase:firebase-analytics'
 
     implementation 'com.github.tbruyelle:rxpermissions:0.10.2'
 
     implementation "com.trello.rxlifecycle2:rxlifecycle-kotlin:$rootProject.rxlifecycle_version"
     implementation "com.trello.rxlifecycle2:rxlifecycle-components:$rootProject.rxlifecycle_version"
 
+
     
 
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
     testImplementation 'junit:junit:4.12'
 
 
@@ -93,8 +98,6 @@ dependencies {
     testImplementation 'org.robolectric:robolectric:3.7.1'
 
 
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
 
     implementation project(':foundationLib')
 }
